build(fuzz): batch 1 container harnesses — blast + pe; fix pe reader overflow (#116) (#322)
fuzz_blast drives both DCL entry points (raw stream + EA size-prefix
wrapper) with tight-capacity edges; fuzz_pe drives the shared MZ/LE
section reader plus VA translation, with a volatile read of the section
bounds so ASan sees bad pointers. Synthetic seeds per the fixture policy;
PL/MZ/CODE dictionary for fuzz_pe.

fuzz_pe found a real bug within 90 seconds: every offset sum in
pe_code_section was 32-bit, so e_lfanew (or a section raw pointer) near
UINT32_MAX wrapped past the bounds checks and read out of bounds —
reachable from any LAY/FNT/MUS overlay file. All sums are now 64-bit;
tests/test_pe.cpp pins the reproducer shapes plus the va translation
edges. LIB/FNT/LAY specs claim the new harnesses; matrix regenerated.

+++ lib/src/pe.cpp
@@ -14,22 +14,24 @@ CodeSection pe_code_section(const uint8_t* data, size_t size) {
     if (size < 0x40 || data[0] != 'M' || data[1] != 'Z')
         return {nullptr, 0, 0};
     uint32_t pe_off = u32le(data + 0x3C);
-    if (pe_off + 24 + 2 > size)
+    // 64-bit sums throughout: every offset here is attacker-controlled, and
+    // 32-bit arithmetic wraps past the bounds checks (found by fuzz_pe).
+    if ((uint64_t)pe_off + 24 + 2 > size)
         return {nullptr, 0, 0};
     const uint8_t* pe = data + pe_off;
     if (pe[0] != 'P' || pe[1] != 'L' || pe[2] != 0 || pe[3] != 0)
         return {nullptr, 0, 0};
     uint16_t num_sec    = u16le(pe + 6);
     uint16_t opt_hdr_sz = u16le(pe + 20);
-    uint32_t sec_table  = pe_off + 24 + opt_hdr_sz;
+    uint64_t sec_table  = (uint64_t)pe_off + 24 + opt_hdr_sz;
     for (uint16_t i = 0; i < num_sec; ++i) {
-        uint32_t sec_off = sec_table + (uint32_t)i * 40;
+        uint64_t sec_off = sec_table + (uint64_t)i * 40;
         if (sec_off + 40 > size) break;
         const uint8_t* sec = data + sec_off;
         uint32_t virt_addr = u32le(sec + 12);  // section VirtualAddress
         uint32_t raw_sz    = u32le(sec + 16);
         uint32_t raw_ptr   = u32le(sec + 20);
-        if (raw_sz > 0 && raw_ptr + raw_sz <= size)
+        if (raw_sz > 0 && (uint64_t)raw_ptr + raw_sz <= size)
             return {data + raw_ptr, raw_sz, virt_addr};
     }
     return {nullptr, 0, 0};
